Test for In-Flight SeqNos and Partial Acks
The `sequence_numbers_in_flight` method should return the number of sequence
numbers that are in the queue for retransmission. If the sender receives a
partial acknowlegement, it should still count the acknowleged sequence numbers
as in-flight. In other words, it should not naively take the distance between
the current sequence number and the left edge of the window.