Add benchmark for Checkpoint 6
The benchmark sets up the routing table as given on the assignment. It then
adds packets in the receive queue for each of the interfaces, with destination
IP addresses randomly chosen from a small predefined set. It finally measures
how long it takes for the router to route all of those packets.