Trim padding when parsing IPv4 packets
If an Ethernet frame is smaller than 60 bytes, it will be padded to that
length. However, the old code that parsed IPv4 packets assumed that all of the
data after the Ethernet header was part of the packet. This led to any padding
bytes being treated as part of the payload.

ARP messages are unaffected since their parsing code always reads from the
front of the packet and thus implicitly ignores any appended data.

This commit uses the length fields in the IPv4 header to determine the payload
length, ignoring any data present after that.